requirements.txt
# Core Django
Django==5.0.6

# REST framework + JWT
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.4.0

# Database config + PostgreSQL (Render injects DATABASE_URL)
dj-database-url==2.2.0
psycopg2-binary==2.9.9

# Static files (Render)
whitenoise==6.7.0

# Optional: for image/file uploads
Pillow==10.3.0

# Deployment
gunicorn==22.0.0